Fixed Printing value more than the given range.
When thread number is more than range to print, it was printing more
than the range. This was happening because we were not checking the
condition again after taking the lock. When current thread was waiting
to acquire the lock, different thread might have updated the value for
the number and when current thread acquired the lock, as it was not
checking the condition again, it was printing the value more than the
range given.